pyproject.toml
[project]
name = "customer-support-chatbot"
version = "0.1.0"
description = "Customer support chatbot demo built with LangChain and LangSmith"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "Mock Project Team"}
]
dependencies = [
  "langchain==0.3.7",
  "langchain-community==0.3.1",
  "langchain-openai==0.2.2",
  "langsmith>=0.1.125",
  "faiss-cpu>=1.8.0.post1",
  "python-dotenv>=1.0.1",
  "typer>=0.12.4",
  "rich>=13.8.1",
  "docx2txt>=0.8",
  "pypdf>=4.3.1",
  "tiktoken>=0.7.0",
  "python-docx>=1.2.0",
  "fpdf>=1.7.2",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = ["tests"]


